refactor out builder class
